CHROMIUM: iwl7000: mac80211: Disable U-APSD if connected to a non-HT AP
This concerns a buggy behavior of iPhone 4/4s when SoftAP, that advertise
themselves falsely as supporting U-APSD.  However, they don't raise the
TIM in its Beacon even if it has data for us, therefore traffic is stuck.
As a workaround, use the fact that those devices does not advertise HT
capabilities, and disable U-APSD if connected to a non-HT AP.
Our previous implementation assumed that if it doesn't have ht_capa
it can't be a SISO iphone.  Evidently this is not true, as non-HT implies
that the AP does not support 11n, hence does not support MIMO.

fixes:I2ea5d8109e0e5ae8b0a24e3f78ff5d0b500022ea

Signed-off-by: Avri Altman <avri.altman@intel.com>
iwl7000-tree: 12daf3f3909457ff4b38d4ad78aab519d5922205