CHROMIUM: tpm: issue TPM_Savestate proactively to prevent suspend delays
(This patch has no effect unless the module parameter
tpm.tpm_early_savestate is set to 1.)

Some TPM models can throttle the TPM_Savestate command under certain
conditions, to prevent wear.  This results in suspend-time increases
which, if ignored, may hide other regressions during automated tests.
It can also interfere with the suspend and make it fail.

This problem is made more acute by the over-throttling behavior
of a subset of those models (which seems to be an error).  In
those models the throttling can be seen even when TPM_Savestate
commands are issued far apart from each other.

The workaround suggested in the bug discussion is to issue a TPM_Savestate
command within some time interval after a command which may change the
volatile state of the TPM.  A TPM_Savestate following such command
requires the TPM to actually write data to the nvram, which
may trigger the throttling.  (If the TPM state has not changed
since the last TPM_Savestate, the TPM doesn't need to save it
again.)

We would also like to avoid issuing unnecessary savestate commands.
If NO commands are issued following the last TPM_Savestate, it
should be OK in general to skip the savestate at suspend.
Unfortunately at this time we rely on this command to help decrement
the dictionary attack counter on some systems, so we leave this
behavior configurable via a module parameter.

BUG=chrome-os-partner:26289
TEST=boot, suspend, resume
Signed-off-by: Luigi Semenzato <semenzato@chromium.org>

Change-Id: Id062df884d7f37d017041f68af75dd03c12a0b26
Reviewed-on: https://chromium-review.googlesource.com/277726
Reviewed-by: Derek Basehore <dbasehore@chromium.org>
Commit-Queue: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>