CHROMIUM: [media] s5p-mfc: Prevent issuing commands to dead hardware
When MFC encounters a critical error, the driver should not issue new
commands to it anymore, but instead fail any ongoing operations and
completely reinitialize the hardware.

This patch adds a hw_error flag to struct s5p_mfc_dev, which is set
on critical hardware error and cleared on deinitialization, to make sure
that new commands will be issued only after next initialization.

BUG=chromium:495960
TEST=youtube;apprtc

Change-Id: Ia6c08daf9def48781459c6f78cc9ce8b55eed758
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/281847
Reviewed-by: Pawel Osciak <posciak@chromium.org>