CHROMIUM: iwl7000: pcie: save and restore the pci device data locally
When we load the iwlwifi module, we briefly start the pci device and
then stop it again, leaving it in D3hot state.  If the platform
suspends before the interface is brought up, the pci bus driver tries
to save and restore the state, but something goes wrong and we can't
access the device anymore.  The bus driver seems to mess with the
device state data, for instance changing the state from D3hot to D0.

To work around this problem, we can store the device state data
locally in the trans_pci structure so we can restore it properly the
next time we want to start it.

Additionally, make sure the device is powered on when freeing the
trans_pcie struct (i.e. on module unload) so that the PCI functions we
call don't fail, we free the saved_state data and leave the device in
as close a state as it was before the module was loaded in the first
place.

Signed-off-by: Luciano Coelho <luciano.coelho@intel.com>
iwl7000-tree: 8cfa4c584bc1729060a6e93ae0e8f3c0e48a40df