CHROMIUM: iwl7000: mvm: remove useless debug message from RX
This message is useless - it's in the good case that always
happens so enabling it doesn't really help. Just remove it.
There are other ways to debug this (e.g. tracing) so there's
no need to add a message in the bad case.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: a963ba8d32511efebb37d1c47d2f2d8b287509b8