CHROMIUM: iwl7000: mac80211: document status.freq restrictions
It's not always necessary to set the status.freq field, for example
when this would be an expensive calculation. It must be set for all
management frames (as they might be reported to userspace), but for
data frames it's not really required. Document this.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: f31d1b2d571d54fe3a4f88fc550aaeff63e44f09