CHROMIUM: iwl7000: chromeOS: make mac80211-bp.h compatible with both cfg80211 versions
Instead of patching the file at generation time, we can rely on the
CONFIG_WIRELESS_38 symbol to determine at build time which version
of cfg80211 the driver is built against.

This removes the single line that was different for drivers against
cfg80211 3.8 and 3.18, allowing replacing one copy of the code with
a symlink in the 3.18 kernel tree.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 756361154a9445f0131e6c84f7073da6c103844e