CHROMIUM: mwifiex: increase number of probes
Sometimes in a busy environment, AP missed to respond probe
request from device. We will increase number for probes sent
by device to increase the chance of finding the APs present
in the environment.

BUG=chromium:514831
TEST=With all 3 patches in this series applied, ran various
stress tests to find given APs in busy environment(100+ APs)
and confirmed scan performance is improved.

Change-Id: Ia662bd9bba04558fad14467fba5ed7deb844af27
Reviewed-on: https://chromium-review.googlesource.com/303986
Reviewed-by: Sameer Nanda <snanda@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Signed-off-by: Amitkumar Karwar <akarwar@marvell.com>
Reviewed-on: https://chromium-review.googlesource.com/305073
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>