CHROMIUM: [media] s5p-mfc: Fix encoder crop calculation
Current crop calculation code does not round down top and left positions
correctly and does not properly check if crop window stays inside image
boundaries. Moreover. it looks like firmware will crash if crop height
is not a multiple of two.

This patch rewrites the code that performs the adjustment of crop window
so that it always stays inside the image boundaries, covers at least all
of the originally specified part of image and has an even height.

BUG=chromium:550602
TEST=Screen share of window with various odd sizes

Change-Id: Ie497748c88aad88882c80b59615a87898459d3fe
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/312912
Reviewed-by: Pawel Osciak <posciak@chromium.org>