CHROMIUM: config: Turn on CONFIG_COMPACTION
This is the equivalent to:
  echo 'CONFIG_COMPACTION=y' >> chromeos/config/base.config
  ./chromeos/scripts/kernelconfig oldconfig

...atop an already normalized config.

Note that compaction has been on by default in the upstream kernel since
2012.  In commit 05106e6a54ae ("mm: enable CONFIG_COMPACTION by
default") from October of 2012 you can see that the argument was:

    Now that lumpy reclaim has been removed, compaction is the only way
    left to free up contiguous memory areas.  It is time to just enable
    CONFIG_COMPACTION by default.

Basically we're hamstringing ourselves for no reason by leaving
CONFIG_COMPACTION off.  Though I'm no expert in compaction, the
description of it (and the fact that it's on by default upstream) makes
me think that if something isn't broken by virtual memory it won't be
broken by compaction.

BUG=chrome-os-partner:45689
TEST=Use device under memory pressure.  No problems.

Change-Id: Idc96bc142bff69024b446d7eba4de53c695d3e96
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/324819
Reviewed-by: Sameer Nanda <snanda@chromium.org>