CHROMIUM: drm/udl: remove drm_device_is_unplugged and related code
v1: Remove the general drm_device_is_unplugged() checks, and move the unplugged
flag handling logic into drm/udl. In general we want to keep driver-specific
logic out of common drm code.

v2: Based on discussion with Stephane and David, drop most of the unplugged
flag handling logic in drm/udl except for udl_detect() and udl_fb_open().
The intention is to treat the device removal as a connector-unplug, and kep
the UDL device fully functional.

v3: Based on feedback from David, entirely drop the "unplugged" flag and all
related code. There is no need to check for the unplugged flag as the existing
udl_usb_disconnect() behavior already ensures the controller is removed, and
all code paths that uses the usb-device are not reachable after unplug.

When a UDL monitor is unplugged, we need to still treat it as a fully
functional device which just happens to have its connector unplugged.
This allows user-space to properly deallocate fbs and dumb buffers
before closing the device.

This drops the "unplugged" flag hack, which puts the device in a
non-functional state after USB unplug and rejects most operations on
the device such as ioctls with error -ENODEV.

Signed-off-by: Haixia Shi <hshi@chromium.org>
Reviewed-by: Stphane Marchesin <marcheu@chromium.org>

BUG=chromium:583759
BUG=chromium:583508
TEST=verify kernel WARN_ON and use-after-free bugs are fixed

Change-Id: I6ab66bcb45a5adc7680c62b91fd48676e658a896
Reviewed-on: https://chromium-review.googlesource.com/325866
Commit-Ready: Haixia Shi <hshi@chromium.org>
Tested-by: Haixia Shi <hshi@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>